_archived/sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mao-testing"
version = "0.1.0"
description = "SDK for MAO Testing Platform - Multi-Agent Orchestration failure detection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.0.1"]
autogen = ["pyautogen>=0.2.0"]
crewai = ["crewai>=0.1.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "respx>=0.20.0",
]
all = [
    "mao-testing[langgraph,autogen,crewai,dev]",
]

[project.urls]
Homepage = "https://github.com/mao-testing/sdk"
Documentation = "https://docs.mao-testing.com"

[tool.setuptools.packages.find]
where = ["."]
include = ["mao_testing*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
